pyproject.toml
[project]
name = "defi-daily-snapshot"
version = "0.1.0"
description = "DefiLlama-powered daily DeFi snapshot bot for Discord + X"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tenacity>=8.0",
    "tweepy>=4.14",
    "tzdata>=2024.1; sys_platform == 'win32'",
    "anthropic>=0.45",
    "google-genai>=1.0",
    "Pillow>=10.0",
    "pilmoji>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"
